requirements.txt
# Python dependencies for the books build system
PyYAML>=6.0
loguru>=0.7.0
